Suport proto_strip_import_prefix
The proto module from `gazelle` allows to strip prefixes directory.

Which is handy if your proto files are in `/proto/foo` but the package
is actually just `foo`.

The problem is that `protoc` don't support that. We aim at changing it
so that `gazelle` can generate also `protoc` related rules.